Fixed out of memory crash
When blocks are being processed and difficulty target function in
AbstractblockChain is run the new devcoin algorithm expects a few
hundred blocks in history to be able to find future difficulty through
averaging, however bitcoin didn't expect this as the last checkpoint is
the only one valid at first, so there is a crash. Instead I wait until
transactions are given to me (after fastCatchup is over) which assumes
that you have downloaded atleast a few weeks of data.